requirements.txt
aiogram==3.2.0
supabase==2.0.3
gotrue==1.3.0
httpx==0.24.1
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1
postgrest==0.13.0
realtime==1.0.2
openpyxl==3.1.2
pandas==2.1.4
